requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
anthropic>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
python-multipart>=0.0.6
e2b>=0.12.0
websockets>=11.0.0